[Q20-13] - addressed review comments - moved the aspect class to core and replaced its config annotation with component
